perf(sqlite): reuse cached statements for schema reads
